+++ SwiftCrypt/Ciphers/Caesar.swift
@@ -9,67 +9,16 @@
 import Foundation
 
 /// The Ceasar Cipher.
-class Caesar {
-    /// A set of rules to determine how to shift each Latin `Character` in a `String`.
-    private struct Shifter {
-        let shift: (Int, Int) -> Int
-        let wrap: (Int, Int) -> Int
-        let uppercaseBoundExceeded: (Int) -> Bool
-        let lowercaseBoundExceeded: (Int) -> Bool
-
-        /// Shifts the `Character` if it is ASCII the appropriate count.
-        ///
-        /// - Parameters:
-        ///   - input: The `Character` to be shifted.
-        ///   - distance: The distance to shift the `Character` input.
-        /// - Returns: The shifted `Character`.
-        func shift(input: Character, distance: Int) -> Character {
-            guard input.isAsciiLetter else { return input }
-
-            var av = shift(input.asciiValue!, distance)
-
-            if (input.isUppercaseAsciiLetter && uppercaseBoundExceeded(av)) ||
-                (input.isLowercaseAsciiLetter && lowercaseBoundExceeded(av)) {
-                    av = wrap(av, 26)
-            }
-
-            return av.charValue!
-        }
-
-        /// Shifts every `Character` that is ASCII in the input `String` the appropriate direction and count.
-        ///
-        /// - Complexity: O(n) where n = input length
-        /// ---
-        /// Example:
-        ///
-        ///     shift("Hello World", withKey: "5", usingShifter: shifter)
-        /// ---
-        /// - Parameters:
-        ///   - input: The `String` to be shifted.
-        ///   - distance: The distance to shift each `Character` in the `String` input.
-        /// - Returns: The shifted input.
-        func shift(input: String, distance: Int) -> String {
-            var output = ""
-
-            for c in input {
-                let shiftedLetter = shift(input: c, distance: distance)
-
-                output += String(shiftedLetter)
-            }
-
-            return output
-        }
-    }
-
-    private static let leftShifter = Shifter(shift: -,
+struct Caesar {
+    private static let leftShifter: Shifter = Shifter(shift: -,
                                       wrap: +,
-                                      uppercaseBoundExceeded: { $0 < ("A" as Character).asciiValue! },
-                                      lowercaseBoundExceeded: { $0 < ("a" as Character).asciiValue! })
+                                      uppercaseBoundExceeded: { $0 < ("A" as Character).decimalValue! },
+                                      lowercaseBoundExceeded: { $0 < ("a" as Character).decimalValue! })
 
-    private static let rightShifter = Shifter(shift: +,
+    private static let rightShifter: Shifter = Shifter(shift: +,
                                        wrap: -,
-                                       uppercaseBoundExceeded: { $0 > ("Z" as Character).asciiValue! },
-                                       lowercaseBoundExceeded: { $0 > ("z" as Character).asciiValue! })
+                                       uppercaseBoundExceeded: { $0 > ("Z" as Character).decimalValue! },
+                                       lowercaseBoundExceeded: { $0 > ("z" as Character).decimalValue! })
 
     private static let invalidCaesarKey = "The given key is invalid and must be a number from 1 to 25"
 }