+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
+## [0.1.3] - 2018-05-12
+### Changed
+- Fixed bugs in Int extension test
+- Fixed bugs in Int extension
+- Made Caesar a struct
+- Abstracted out Shifter into a reusable struct
+
 ## [0.1.2] - 2018-03-31
 ### Changed
 - Fixed bugs for the Character extension